refactoring/testing the 3D tiles extensions
